requirements.txt
annotated-types
blis
catalogue
certifi
charset-normalizer
click
cloudpathlib
confection
curated-tokenizers
curated-transformers
cymem
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
filelock
fsspec
huggingface-hub
idna
Jinja2
langcodes
language_data
marisa-trie
markdown-it-py
MarkupSafe
mdurl
mpmath
murmurhash
networkx
numpy
packaging
preshed
pydantic
pydantic_core
Pygments
PyYAML
regex
requests
rich
safetensors
shellingham
smart-open
spacy
spacy-curated-transformers
spacy-legacy
spacy-loggers
srsly
sympy
thinc
tokenizers
torch
tqdm
transformers
typer
typing_extensions
urllib3
wasabi
weasel
wrapt
redis
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
anyio
ariadne-codegen
attrs
autoflake
backoff
beautifulsoup4
black
blinker
cachetools
cffi
cloudevents
colorama
colorclass
compressed_rtf
cryptography
deprecation
Distance
dnspython
dnstwist
easygui
ebcdic
email_validator
extract-msg
fastapi
Flask
frozenlist
functions-framework
geoip2
gql
graphql-core
gunicorn
h11
httpcore
httpx
ijson
iniconfig
isort
itsdangerous
lark
lxml
maxminddb
msoffcrypto-tool
multidict
mypy
mypy-extensions
netaddr
olefile
oletools
orjson
pathspec
pcodedmp
platformdirs
pluggy
ppdeep
propcache
psutil
py-cpuinfo
py-tlsh
pycparser
pyflakes
PyJWT
pyparsing
pytest
python-dateutil
python-dotenv
python-multipart
pytz
querycontacts
red-black-tree-mod
requests-file
requests-toolbelt
RTFDE
ruff
setuptools
six
sniffio
soupsieve
starlette
tenacity
tld
tldextract
toml
twilio
tzlocal
ujson
uvicorn[standard]
vt-py
watchdog
Werkzeug
wheel
yarl